Move getNetworkInterface to run_common.go
